Add suse support for the snaps build
suse snaps rpm build and test happens inside a docker container
This patch requires the following patch to be merged:

https://git.opendaylight.org/gerrit/#/c/68114/

It reuses most of the building and testing scripts. Those scripts
for epel will be executed inside a VM (as it was done before) and
for suse, those will be executed inside a docker container.

It avoids duplicating the packaging-build-rpm-snap job using the
variable {distro} to differentiate between suse and epel. However,
to reuse most building and testing scripts, we inject the DISTRO
variable into the scripts to detect if we are inside the VM or the
container. When building a suse rpm, all scripts will be executed
twice and thus if we detect that it is a suse rpm and we are inside
a VM, it will exit the script

+++ jjb/packaging/build-rpm-snap.sh
@@ -6,6 +6,20 @@
 #   -o pipefail: Fail on errors in scripts this calls, give stacktrace
 set -ex -o pipefail
 
+#-------------------------------------------------------------------------------
+# Exit if suse and in a VM
+#-------------------------------------------------------------------------------
+# Jenkins template and scripts are shared between suse and red hat to build and
+# test the rpms. However, all the suse processing is done in a container whereas
+# redhat processing is done in a VM. We should exit if we detect that this
+# script is going to build a suse rpm inside a VM. DISTRO variable only exists
+# when the script is executed in the VM.
+#-------------------------------------------------------------------------------
+if [ "$DISTRO" == "suse" ]; then
+  echo "We are in a VM, nothing to do for suse"
+  exit 0
+fi
+
 # Install required packages
 virtualenv rpm_build
 # shellcheck disable=SC1091
@@ -38,11 +52,19 @@ if [ "$SILO" == "sandbox" ]; then
   # TODO: Host RPMs on Jenkins temporarily
   echo "Not uploading RPMs to Nexus because running in sandbox"
 elif  [ "$SILO" == "releng" ]; then
-  # Move RPMs (SRPM and noarch) to dir of files that will be uploaded to Nexus
-  UPLOAD_FILES_PATH="$WORKSPACE/upload_files"
-  mkdir -p "$UPLOAD_FILES_PATH"
-  cp "/home/$USER/rpmbuild/RPMS/noarch/"*.rpm "$_"
-  cp "/home/$USER/rpmbuild/SRPMS/"*.rpm "$_"
+  if [ -f /usr/bin/yum ]; then
+    # Move RPMs (SRPM and noarch) to dir of files that will be uploaded to Nexus
+    UPLOAD_FILES_PATH="$WORKSPACE/upload_files"
+    mkdir -p "$UPLOAD_FILES_PATH"
+    cp "/home/$USER/rpmbuild/RPMS/noarch/"*.rpm "$_"
+    cp "/home/$USER/rpmbuild/SRPMS/"*.rpm "$_"
+  elif [ -f /usr/bin/zypper ]; then
+    # Move RPMs (SRPM and noarch) to dir of files that will be uploaded to Nexus
+    UPLOAD_FILES_PATH="$WORKSPACE/upload_files"
+    mkdir -p "$UPLOAD_FILES_PATH"
+    cp "/root/rpmbuild/RPMS/noarch/"*.rpm "$_"
+    cp "/root/rpmbuild/SRPMS/"*.rpm "$_"
+  fi
 else
   echo "Unknown Jenkins silo: $SILO"
   exit 1
